RLSB3_requirements.txt
absl-py==1.3.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cloudpickle==2.2.0
contourpy==1.0.5
cycler==0.11.0
fenics-dijitso @ file:///home/conda/feedstock_root/build_artifacts/fenics-pkgs_1662986692496/work/dijitso
fenics-dolfin @ file:///home/conda/feedstock_root/build_artifacts/fenics-pkgs_1662986692496/work/dolfin/python
fenics-ffc @ file:///home/conda/feedstock_root/build_artifacts/fenics-pkgs_1662986692496/work/ffc
fenics-fiat @ file:///home/conda/feedstock_root/build_artifacts/fenics-pkgs_1662986692496/work/fiat
fenics-ufl @ file:///home/conda/feedstock_root/build_artifacts/fenics-pkgs_1662986692496/work/ufl
fonttools==4.37.4
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1641732907644/work
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
gym==0.21.0
idna==3.4
importlib-metadata==4.13.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
mpi4py @ file:///home/conda/feedstock_root/build_artifacts/mpi4py_1660326282916/work
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1612895720168/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1662888869821/work
oauthlib==3.2.1
packaging==21.3
pandas==1.5.0
PeakUtils @ file:///home/conda/feedstock_root/build_artifacts/peakutils_1659963035671/work
petsc4py @ file:///home/conda/feedstock_root/build_artifacts/petsc4py_1659515074188/work
pickle5==0.0.11
Pillow==9.2.0
pkgconfig @ file:///home/conda/feedstock_root/build_artifacts/pkgconfig_1648797873901/work
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11 @ file:///home/conda/feedstock_root/build_artifacts/pybind11-split_1657939773764/work
pybind11-global @ file:///home/conda/feedstock_root/build_artifacts/pybind11-split_1657939773764/work
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sb3-contrib==1.6.2
scipy==1.9.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
slepc4py @ file:///home/conda/feedstock_root/build_artifacts/slepc4py_1660112752507/work
stable-baselines3==1.6.2
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1664580364308/work
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.9.0
